refactor: claude refactor state into store
and other minor commentns
